api/document-api/requirements.txt
accelerate==1.7.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aioitertools==0.12.0
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
billiard==4.2.1
blis==1.2.1
boto3==1.37.3
botocore==1.37.3
botocore-stubs==1.38.26
catalogue==2.0.10
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.1
cohere==5.2.6
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.3
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
elastic-apm==6.9.1
elasticsearch==7.13.0
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastapi==0.115.0
fastavro==1.11.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
flower==2.0.1
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.5.1
greenlet==3.2.2
grpc-interceptor==0.15.4
grpcio==1.72.1
grpcio-tools==1.72.1
h11==0.16.0
hf-xet==1.1.2
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
humanfriendly==10.0
humanize==4.12.3
idna==3.10
importlib_metadata==8.6.1
invoke==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpath-python==1.0.6
kiwisolver==1.4.8
kombu==5.5.3
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
libmagic==1.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.4.4
murmurhash==1.0.13
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
olefile==0.47
onnx==1.18.0
onnxruntime==1.22.0
openai==1.82.1
opencv-python==4.11.0.86
openparse==0.7.0
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20240706
pi_heif==1.1.1
pikepdf==9.0.0
pillow==12.0.0
pillow-heif==1.1.1
pinecone-text==0.9.0
pluginbase==1.0.1
preshed==3.0.10
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.4.0
PyMuPDF==1.26.0
pyparsing==3.2.3
pypdf==5.5.0
pypdfium2==4.30.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==4.3.1
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==14.0.0
s3transfer==0.11.3
safetensors==0.5.3
scipy==1.15.3
semantic-router==0.0.72
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.5.1
starlette==0.38.6
sympy==1.14.0
tabulate==0.9.0
TExtractor==0.1.2
thinc==8.3.4
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.15.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.39.3
typer==0.16.0
types-aiobotocore==2.22.0
types-aiobotocore-s3==2.22.0
types-aiobotocore-sts==2.22.0
types-awscrt==0.27.2
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.28.1
unstructured-inference==1.0.2
unstructured.pytesseract==0.3.15
urllib3==1.26.20
uvicorn==0.30.6
vine==5.1.0
Wand==0.6.13
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
wget==3.2
wrapt==1.17.2
xlrd==2.0.1
yarl==1.20.0
zipp==3.22.0
